requirements.txt
torch>=2.2
torchvision>=0.17
tqdm
numpy
opencv-python
Pillow
PyYAML
av